+++ example/override/sfmlboxinstanceinfo.cpp
@@ -8,7 +8,7 @@ namespace SpriterEngine
 	SfmlBoxInstanceInfo::SfmlBoxInstanceInfo(point initialSize, sf::RenderWindow *validRenderWindow) :
 		BoxInstanceInfo(initialSize),
 		renderWindow(validRenderWindow),
-		rectangle(sf::Vector2f(initialSize.x, initialSize.y))
+		rectangle(sf::Vector2f(static_cast<float>(initialSize.x), static_cast<float>(initialSize.y)))
 	{
 	}
 
@@ -16,10 +16,10 @@ namespace SpriterEngine
 	{
 		if (Settings::renderDebugBoxes)
 		{
-			rectangle.setPosition(getPosition().x, getPosition().y);
-			rectangle.setRotation(toDegrees(getAngle()));
-			rectangle.setScale(getScale().x, getScale().y);
-			rectangle.setOrigin(getPivot().x*getSize().x, getPivot().y*getSize().y);
+			rectangle.setPosition(static_cast<float>(getPosition().x), static_cast<float>(getPosition().y));
+			rectangle.setRotation(static_cast<float>(toDegrees(getAngle())));
+			rectangle.setScale(static_cast<float>(getScale().x), static_cast<float>(getScale().y));
+			rectangle.setOrigin(static_cast<float>(getPivot().x * getSize().x), static_cast<float>(getPivot().y * getSize().y));
 			renderWindow->draw(rectangle);
 		}
 	}
