+++ spriterengine/loading/spriterdocumentloader.cpp
@@ -1354,7 +1354,7 @@ namespace SpriterEngine
 				if (att->getName() == "parent")
 				{
 					int parentRefIndex = att->getIntValue();
-					if (parentRefIndex < refObjectIds.size())
+					if (parentRefIndex < static_cast<int>(refObjectIds.size()))
 					{
 						parentObjectId = refObjectIds.at(parentRefIndex);
 					}
@@ -1370,7 +1370,7 @@ namespace SpriterEngine
 					objectId = animation->getObjectIdFromTimelineIndex(timelineIndex);
 					refObjectIds.push_back(objectId);
 
-					if (timelineIndex >= 0 && timelineIndex < timelineRedunantKeyedVector->size())
+					if (timelineIndex >= 0 && timelineIndex < static_cast<int>(timelineRedunantKeyedVector->size()))
 					{
 						if (timelineRedunantKeyedVector->at(timelineIndex))
 						{
@@ -1424,7 +1424,7 @@ namespace SpriterEngine
 				if (it != spriteKeyFileInfoMap->end())
 				{
 					SpriteKeyFileInfoVector *spriteKeyInfoVector = &(*it).second;
-					if (keyIndex < spriteKeyInfoVector->size())
+					if (keyIndex < static_cast<int>(spriteKeyInfoVector->size()))
 					{
 						SpriteKeyFileInfo *spriteKeyInfo = &spriteKeyInfoVector->at(keyIndex);
 						mainlineKey->pushBackZOrderRef(new SpriteRef(objectId, parentObjectId, timelineKey, spriteKeyInfo->fileIndex, spriteKeyInfo->useDefaultPivot));
@@ -1441,7 +1441,7 @@ namespace SpriterEngine
 					if (it != subEntityKeyInfoMap->end())
 					{
 						SubEntityKeyInfoVector *subEntityKeyInfoVector = &(*it).second;
-						if (keyIndex < subEntityKeyInfoVector->size())
+						if (keyIndex < static_cast<int>(subEntityKeyInfoVector->size()))
 						{
 							SubEntityKeyInfo *subEntityKeyInfo = &subEntityKeyInfoVector->at(keyIndex);
 							mainlineKey->pushBackZOrderRef(new EntityRef(objectId, parentObjectId, timelineKey, subEntityKeyInfo->entityId, subEntityKeyInfo->animationIndex));
